move use on demand utils
